django_weather_reminder/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-templated-mail==1.1.1
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==19.3.1
idna==3.4
iniconfig==2.0.0
kombu==5.3.1
oauthlib==3.2.2
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
flower ==1.2.0
gunicorn==21.2.0